Entities: Create entity which maintains pos, color and inturn draw
Inturn create a set of entities and inturn update them on screen,
based on a simple logic for now.

NOTE: This wont work, as player is moved out of Players vector in
the loop, will be changing in next commit. This is just as a memory
feedback to self, for future.