Rcg: bit asymmetrical field boundry, to try match asymmetry seen
To rcg files, which I tried to playback using this logic, seemed
to show the game playback being bit offset, this commit tries to
overcome that offset issue to an extent, in a blind way.

TODO: Need to check max seen x,y from these rcg files to try
identify the actual boundry in a better way. Also need to look
at the source and documentation around robocup soccer sim later,
to cross check same and update if required.