AliceBackInNormalLand:Ooops: Rect offset was origin of asymmetry
As now I do a fill rect taking position as mid point of rect, so
the original asymmetry seen is fixed, so revert the Rcg boundry
asymmetric adjustment back to normal.

Also this means no need to look at source or doc of rcss later, as
things seem to be fine.

WHile at it, also make the ball use the new fill_rect_mid, this
seems to make the ball sit perfectly wrt the boundry line, during
kick_in, so assuming things are symmetric as it should be, things
should be fine vertically. Horizontally the boundry condition (ie
the left and right goal post etal), I need to check once later.