SecondsPerRecord-FramesPerSecond synced and used more appropriately
Rather each playdata type can specify the seconds between the records
it contains and inturn provides (it can even be a fraction of a second),

Inturn any changes to fps used by playbackpgnd is notified to currently
active playdata instance, So that it can handle NextFrameIsRecordReady
properly.

The same two inturn are used, when deciding the inframes to pass wrt
update call, if interpolated movements is being used.