Main: Show also the actual rough frames per sec
NOTE: I dont account for processing overhead currently, before
deciding to sleep for time corresponding to fps. Also this is
not a realtime scheduled process, so sleep wont return in time
most of the time, so actual fps will be different from set fps,
this shows both at the center top.