Gui: Sleep only for remaining time wrt a frame
Track time consumed wrt processing related to handling things reqd
for a frame, and inturn only sleep for the remaining time wrt the
frame.

This helps try make the actualfps achieved more closer to requested
fps for a wider range of fps for release as well as debug builds.

Program version updated.