ProcActions: Merge kicks and tackles into a single actions vec
The scores have also been adjusted to ensure that a fair score
is given wrt most sequences.

NOTE: If one where to retain the seperate vectors wrt kicks and
tackles, then one would have had to introduce a last action time
into ActionsInfo struct. And inturn when ever the same side and
same player is noticed wrt a handle_kick/tackle logic, one would
have had to check this saved last action time and then decide,
whether to account the current action or not and so on...