ProcActions: Pass position to the handle function
As the position may be before or after the action info, depending
on how the playdata source pushes the data in, so collate all info
and inturn call the new handle_action function at the end.

This also helps with not requiring to call for updating position
info in all cases to the ActionsInfo logic, in future, when it
will also track total movement wrt each player.