Goal+: Find immidiately prev kick action
Even if something else inbtw, bcas there could be tackle or so
in between in some cases.

TODO: If there is a catch inbetween, then ideally it is not fine
for a goal, this is not taken care of currently.

Misc: Panthrand, Ammu

NOTE: Even though goal time ball position is added to actions list,
thorught the goal action, the virtual ball wont move to that location
wrt rcg, bcas the game time/timecounter doesnt change after goal,
till next kick-off.