ProcActions: Fix rev for loops, Switch handle_action version used
Rather as the range operation by default doesnt include the ending
value, so the index generated by iterating through a range of len
of a vector will not try access the vector using index = len value.
This inturn also holds true when rev() is applied to the range based
iterator. So avoid the -1 that I had wrongly applied wrt this situa\
tion.

Also enable the new and partial handel action logic by renaming the
old and the new functions.