ProcActions: Adjust player scoring wrt goal in the goal chain
Adjust player id to account for the new goal chain based scoring.

Normally a auto infered goal event doesnt contain playerid of the
responsible player wrt the goal, and the logic is required to set
the responisble player in the goal actiondata. However if the goal
comes as part of a manual goal record, then may be someone may
not follow the normal convention and instead specify the player
responsible for the goal in the action data itself, this was
accounted in the goal scoring logic.

HOWEVER with the new goal chain traversal based flow to allow more
players who also would have helped with achieving the goal, to be
rewarded, this overriding of prevactd playerid with curactd pid,
wont work beyond the 1st lookback, this is accounted now.

NOTE: THis still makes the assumption that a valid playerid if any
mentioned in the goal actiondata will correspond to the playerid
in the immidiate previous actiondata wrt that side. Ie in the non
self goal case.