ProcActions: Look4 goalChain based scoring during self goal also
Many a times a self goal, involves the efforts of the players of the
side which got the goal also, to a great extent. This commit tries
to account for the same.

However if a goal losing side player is noticed in the chain beyond
the immidiate previous action before the goal action, then the chain
traversal is terminated.