ProcActions:HandleGoal: GoalChain less otherside penalising
The otherside player is aleady penalised for losing control of the
ball, before hand, so dont penalise them too much again during the
goal chain traversing. Also the penalising occurs only for the 1st
otherside player in the chain (beyond any immidiate self goal one).

However come to think of it, in a way, the otherside players involved
need to be penalised, as they should have been defending againt the
opposite side trying to get the goal. So some amount of penalising
is needed. TODO: And maybe even others beyond the initial 1 or 2
that will be penalised currently?!?