DrawPrimitivies+GEntity: Allow extra graphics wrt GEntities
Add support for some enum based draw primitives to be specified wrt
a GEntity if required.

Also add a Enum to represent some of the draw primitives supported
by sdlx. Add a dummy test func in testlib so that I dont have to
put allow-dead-code wrt this.

These enums allow a vector of graphic operations to be chained on,
if required.

Make draw call take a mutable reference, so that the object being
drawn can update its internal state.