ProcActions: Avoid CrossCorr based filtering of Plot TVS data
This ensures that changes to scoring is immidiately visible as well
as match the actual value, rather than the filtered value, that would
have got plotted previously (along with last entry(ies) not getting
reflected immidiately, previously).

This makes the plot a little bit less smooth, but here that should
be fine.